+++ bootstrap/cache/packages.php
@@ -1,0 +1,30 @@
+<?php return array(
+    'matthewbdaly/artisan-standalone' =>
+        array(
+            'providers' =>
+                array(
+                    0 => 'Matthewbdaly\\ArtisanStandalone\\Providers\\ServiceProvider',
+                ),
+        ),
+    'nesbot/carbon' =>
+        array(
+            'providers' =>
+                array(
+                    0 => 'Carbon\\Laravel\\ServiceProvider',
+                ),
+        ),
+    'tychovbh/laravel-package-skeleton' =>
+        array(
+            'providers' =>
+                array(
+                    0 => 'Tychovbh\\LaravelPackageSkeleton\\PackageSkeletonServiceProvider',
+                ),
+        ),
+    'tychovbh/laravel-mvc' =>
+        array(
+            'providers' =>
+                array(
+                    0 => 'Tychovbh\\Mvc\\MvcServiceProvider',
+                ),
+        ),
+);