+++ src/Http/Controllers/Laravel/Controller.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Tychovbh\Mvc\Http\Controllers\Laravel;
+
+use App\Http\Controllers\Rest;
+use Illuminate\Routing\Controller as BaseController;
+
+class Controller extends BaseController
+{
+    use Rest;
+}