+++ src/Http/Controllers/Lumen/Controller.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Tychovbh\Mvc\Http\Controllers\Lumen;
+
+use App\Http\Controllers\Rest;
+use Laravel\Lumen\Routing\Controller as BaseController;
+
+class Controller extends BaseController
+{
+    use Rest;
+}