+++ src/Http/Requests/Lumen/FormRequest.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Tychovbh\Mvc\Http\Requests\Lumen;
+
+use Urameshibr\Requests\FormRequest as BaseFormRequest;
+use Tychovbh\Mvc\Http\Requests\RequestSettings;
+
+class FormRequest extends BaseFormRequest
+{
+    use RequestSettings;
+}