+++ src/MvcServiceProvider.php
@@ -3,7 +3,10 @@
 namespace Tychovbh\Mvc;
 
 use Illuminate\Support\ServiceProvider;
-use Tychovbh\Mvc\Console\MvcRepository;
+use Tychovbh\Mvc\Console\Commands\MvcRepository;
+use Tychovbh\Mvc\Console\Commands\MvcController;
+use Tychovbh\Mvc\Console\Commands\MvcRequest;
+use Tychovbh\Mvc\Http\Middleware\ValidateMiddleware;
 
 class MvcServiceProvider extends ServiceProvider
 {
@@ -14,9 +17,26 @@ class MvcServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        $this->commands([
-            MvcRepository::class,
-        ]);
+        if ($this->app->runningInConsole()) {
+            $this->commands([
+                MvcRepository::class,
+                MvcController::class,
+                MvcRequest::class
+            ]);
+        }
+
+        $source = __DIR__ . '/../config/messages.php';
+        if (is_application() === 'laravel') {
+            $this->publishes([$source => config_path('messages.php')], 'laravel-mvc');
+        }
+
+        if (is_application() === 'lumen') {
+            $this->app->configure('messages');
+            $this->app->register(\Urameshibr\Providers\FormRequestServiceProvider::class);
+            $this->app->routeMiddleware([
+                'validate' => ValidateMiddleware::class
+            ]);
+        }
     }
 
     /**