+++ src/files/LaravelController.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace App\Http\Controllers;
+
+use Tychovbh\Mvc\Http\Controllers\Laravel\Controller as BaseController;
+
+class EntityController extends BaseController
+{
+    //
+}