+++ src/files/Repository.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+namespace App\Repositories;
+
+use Tychovbh\Mvc\Repositories\Repository;
+use Tychovbh\Mvc\Repositories\AbstractRepository;
+
+class EntityRepository extends AbstractRepository implements Repository
+{
+    //
+}